requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
Flask==2.1.1